Isolate download link code to another component
